Fix coverage: - No need to compare type when unifying ValueTrack. - myia_hasnext and myia_next always receive expected types, already checked by myia_iter - do not check that _bin_rop_inferrer receive different operand types - InferenceDefinition.type_to_abtract never receive lists/tuples currently. - remove unused symbols from common - move dataclasses from common to test_infer (as they are used only there) - Use A instead of Ex and External - Inline scalar_cast and unsafe_static_cast - Remove unused _check_inference() overload in multitest - Test than we cannot unify dicts with different sizes. - Add inference test for range - Add inference test for list getitem (index and slice) - Remove useless test_multitype
