+++ myia/testing/common.py
@@ -1,6 +1,4 @@
 """Common testing utilities."""
-from dataclasses import dataclass
-
 from ovld import ovld
 
 from myia import basics
@@ -81,26 +79,6 @@ def D(**kwargs):
     return A(kwargs)
 
 
-def Ex(value, t=None):
-    """Abstract external from master branch.
-
-    Just create an abstract value.
-    """
-    if value is data.ANYTHING:
-        return A(object)
-    value_type = _to_abstract(value)
-    if t is not None:
-        assert value_type is t, (value, t, value_type)
-    return value_type
-
-
-def Shp(*dims):
-    """Create an abstract tuple representing a shape."""
-    return data.AbstractStructure(
-        [A(dim) for dim in dims], {"interface": tuple}
-    )
-
-
 def build_node(g, descr, nodeset=set()):
     """Create a node recursively from a tuple of tuples.
 
@@ -145,13 +123,6 @@ def _abstract_sequence(seq_type, *items):
     )
 
 
-def _abstract_sequence_getitem(seq_type):
-    def getitem(self, *items):
-        return _abstract_sequence(seq_type, *items)
-
-    return getitem
-
-
 def tuple_of(*items):
     """Create an abstract tuple."""
     return _abstract_sequence(tuple, *items)
@@ -168,66 +139,5 @@ Bot = Nil = A(None)
 Float = A(float)
 Int = A(int)
 Number = Un(int, float)
-External = Ex
 String = A(str)
 EmptyTuple = A(tuple)
-
-
-@dataclass(frozen=True)
-class Point:
-    """Common dataclass for a 2D point."""
-
-    x: int
-    y: int
-
-    def abs(self):
-        """Compute distance from this point to origin."""
-        return (self.x ** 2 + self.y ** 2) ** 0.5
-
-    @property
-    def absprop(self):
-        """Return abs as a property."""
-        return self.abs()
-
-
-@dataclass(frozen=True)
-class Point3D:
-    """Common dataclass for a 3D point."""
-
-    x: object
-    y: object
-    z: object
-
-    def abs(self):
-        """Compute distance from origin to this point."""
-        return (self.x ** 2 + self.y ** 2 + self.z ** 2) ** 0.5
-
-
-@dataclass(frozen=True)
-class Thing:
-    """Common dataclass to use for tests."""
-
-    contents: object
-
-    def __call__(self):
-        """Overload of call."""
-        return self.contents * 2
-
-
-Thing_f = Thing(1.0)
-Thing_ftup = Thing((1.0, 2.0))
-
-
-@ovld
-def mysum(x):  # noqa: F811
-    return x
-
-
-@ovld
-def mysum(x, y):  # noqa: F811
-    return x + y
-
-
-@ovld
-def mysum(x, y, z):  # noqa: F811
-    return x + y + z
